rename train_hf.py to train.py
